+++ include/nfx/Silicon.h
@@ -48,6 +48,7 @@
 #include "nfx/silicon/gate/SRLatch.h"
 
 // 74xx packages
+#include "nfx/silicon/gate/sn74xxx/LS00.h"
 #include "nfx/silicon/gate/sn74xxx/LS04.h"
 #include "nfx/silicon/gate/sn74xxx/LS08.h"
 #include "nfx/silicon/gate/sn74xxx/LS74.h"